backend/requirements.txt
# Django and Django REST Framework
Django>=5.0.0,<5.1.0
djangorestframework>=3.14.0,<3.15.0

# JWT Authentication
djangorestframework-simplejwt>=5.3.0,<5.4.0

# CORS Headers for Flutter frontend
django-cors-headers>=4.3.0,<4.4.0

# MySQL Database Driver
mysqlclient>=2.2.0,<2.3.0

# Image Processing (for profile images)
Pillow>=10.0.0,<11.0.0

# Python Environment Variables
python-decouple>=3.8,<3.9

# API Documentation (optional but recommended)
drf-yasg>=1.21.0,<1.22.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-django>=4.5.0,<4.6.0
factory-boy>=3.3.0,<3.4.0

# Code Quality
flake8>=6.1.0,<7.0.0
black>=23.0.0,<24.0.0

# Production Server (optional)
gunicorn>=21.2.0,<22.0.0
whitenoise>=6.6.0,<7.0.0
